+++ tests/pipeline/test_config.py
@@ -118,6 +118,10 @@ class TestPreprocConfig:
         assert config.subtract_coro_from_center is False
         assert config.flux_combination_method == "median"
         assert config.frame_types_to_extract == ['FLUX', 'CENTER', 'CORO']
+        assert config.flux_cube_selection_irdis == "auto"
+        assert config.flux_cube_selection_ifs == "all"
+        assert config.flux_saturation_adu == 40000.0
+        assert config.flux_nonlinearity_adu == 30000.0
         assert config.eso_username is None
 
     def test_preproc_config_merge(self):